DC28 Badge case
This case has 2 lids one that slides in and is the most secure the other pops in and sometimes works well but sometimes it does not. so use it as a sign or dogtags or a windshield scraper or what ever you want they are cool! 

this was a highly modified Pwnagotchi case originaly made by aethenelle on thingiverse.