Add files via upload
DCZia Def Con 28 Build-A-Badge case. This file has 2 lids, one is bigger and will slid in to the slot. The other is smaller and will pop in on top! The pop in lid is not as secure as the slide in lid and you can use the other lid as what ever you want like a plaque or dog tag or windshield scraper have fun!

This is a modified pwnagotchi case originally made by aethenelle on thingiverse